requirements.txt
Pillow>=10.0.0
platformdirs>=4.0.0

